backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
onnxruntime==1.19.2
pillow==10.4.0
numpy==2.1.1
python-multipart==0.0.9
